Use dev-master version of drush
Now that https://github.com/drush-ops/drush/pull/1595 has landed, and per the comments in #52, the new "finder" stage for Drush is helpful/necessary to use the copy in the vendor directory. In my case, this works without any other changes or magic; I have a global /etc/drush/drushrc.php file that specifies the path to Drush, but this works inside the web/ directory, as well.

Also removed an unnecessary blank line.