Sync new lines at the end of the files
This patch syncs new lines at the end of the files across the source
code.

According to POSIX [1] a line is a sequence of zero or more
`non- <newline>` characters plus a terminating `<newline>` character.

Although it is not mandatory for all files to have a single final new
line, this patch simplifes and syncs the source code files where
possible to have a single final new line either using a `LF` or `CRLF`
character as used prior to this patch in particular file.

This patch also makes editing simpler in certain editors which
automatically or with `.editorconfig` add final new lines to
files. Some editors even automatically trim end of the files to have
only single new line.

Refs:
[1] http://pubs.opengroup.org/onlinepubs/9699919799/basedefs/V1_chap03.html#tag_03_206