Change createRequiredFiles messages to reflect what occured
The messages displayed in the terminal after composer creates the settings.php file, and the files directory were ambiguous because they may lead users to think that they need to create the file and the directory themselves. Changing 'Create' to 'Created' better reflects the fact that they have already been created.